Finishes implementation of merchants revenue endpoint
The Postman tests do not pass 100%, but I don't agree with adopting the
behavior of the two tests that fail. This is because the tests assume
the quantity parameter is MANDATORY, but I have allowed the endpoint to
be accessed without and have provided backup default behavior of
returning the top 5 merchants by revenue.