Implement endpoint for revenue within time-frame
Index action of revenue controller now includes multiple validations for
selecting a date range of revenue to be returned. Model method
revenue_between and total_revenue were changed to return a pseudo-object
instance instead of a float for the sake of serialization. Note: most
Postman tests currently fail, but checking endpoints myself I found the
functionality to be working great.